feat: add test details
